requirements.txt
cloudscraper==1.2.71
beautifulsoup4==4.12.3
google-generativeai
python-dotenv
